Moved motion and LED queue labels to header bar tags
